+++ src/apps/tools/apps/exports/exports/EditionExport.ts
@@ -3,6 +3,7 @@ import { createQueryBuilder, SelectQueryBuilder } from "typeorm";
 
 import { ContributionType } from "@core/utils";
 import { Param } from "@core/utils/params";
+import { convertAnswers } from "@core/utils/polls";
 
 import Member from "@models/Member";
